requirements.txt
# Core PDF Processing
PyMuPDF>=1.23.0
pdfplumber>=0.10.0
python-docx>=1.0.0
openpyxl>=3.1.0

# OCR and Layout Analysis (Optional)
# pytesseract>=0.3.10
# paddleocr>=2.7.0
Pillow>=10.0.0

# AI/LLM - Claude
anthropic>=0.18.0

# Vector DB and Embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0

# Web UI
streamlit>=1.28.0

# PDF Generation
reportlab>=4.0.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
pyyaml>=6.0.0
